generic blank-state component
