+++ app/react/V2/Components/UI/index.ts
@@ -19,6 +19,7 @@ export { NeedAuthorization } from './NeedAuthorization.js';
 export { ConfirmNavigationModal } from './ConfirmNavigationModal.js';
 export { Truncate } from './Truncate.js';
 export { VerticalDrawer } from './VerticalDrawer.js';
+export { BlankState } from './BlankState.js';
 
 export type { PillColor } from './Pill.js';
 export type { TableProps, TableRow } from './Table/index.js';
