+++ app/react/V2/Routes/Entity/Components/BlankState.tsx
@@ -1,22 +1,8 @@
 /* eslint-disable react/no-multi-comp */
 import React from 'react';
-import type { ReactNode } from 'react';
 import { MagnifyingGlassIcon } from '@heroicons/react/24/solid';
 import { Translate } from '#app/I18N/index.js';
-
-type BlankStateProps = {
-  icon: ReactNode;
-  title: ReactNode;
-  description: ReactNode;
-};
-
-const BlankState = ({ icon, title, description }: BlankStateProps) => (
-  <div className="flex flex-col items-center justify-center h-full w-full text-center border border-dashed border-gray-200 rounded-2xl text-gray-500 gap-2">
-    <p className="font-semibold text-lg">{title}</p>
-    {icon}
-    <p className="text-sm text-gray-400 w-4/6">{description}</p>
-  </div>
-);
+import { BlankState } from '#app/V2/Components/UI/index.js';
 
 const NoSearch = () => (
   <BlankState
@@ -38,4 +24,4 @@ const NoResults = () => (
   />
 );
 
-export { BlankState, NoSearch, NoResults };
+export { NoSearch, NoResults };
