+++ app/react/V2/Routes/Entity/Components/ReferencesPanel/SelectTargetStep.tsx
@@ -4,9 +4,8 @@ import { t, Translate } from '#app/I18N/index.js';
 import { ClientRelationshipType } from '#app/apiResponseTypes.js';
 import { Entity } from '#V2/domain/index.js';
 import { FileType } from '#shared/types/fileType.js';
-import { Card } from '#V2/Components/UI/index.js';
+import { Card, BlankState } from '#V2/Components/UI/index.js';
 import { InputField, Checkbox } from '#V2/Components/Forms/index.js';
-import { BlankState } from '../BlankState.js';
 import { EntitySearchResult } from './EntitySearchResult.js';
 import { ReferenceMode } from './useCreateReferenceState.js';
