+++ app/react/V2/Routes/Entity/Components/ToC/ToCPanel.tsx
@@ -10,9 +10,8 @@ import { update as updateFile } from '#V2/api/files/index.js';
 import { FileType } from '#shared/types/fileType.js';
 import { FetchResponseError } from '#shared/JSONRequest.js';
 import { Button } from '#V2/Components/UI/Button.js';
-import { NeedAuthorization } from '#V2/Components/UI/index.js';
+import { NeedAuthorization, BlankState } from '#V2/Components/UI/index.js';
 import { notificationAtom } from '#V2/atoms/index.js';
-import { BlankState } from '../BlankState.js';
 import { ToC, type ProcessedTocEntry, sortTocEntries } from './ToC.js';
 import { entityLoaderCache } from '../../EntityLoaderCache.js';
 import { useToc, useTocActions } from './tocAtom.js';
